fix: opt transforms step over whole multi-byte characters (#107) (#122)
opt_delet() walked a line assuming a kanji is two bytes -- the Shift-JIS
width -- while isjpspace() already matched all three bytes of U+3000 in
UTF-8. fopt=2 ('replace') overwrote the first two bytes of it with ASCII
spaces and left the third in the file as a stray continuation byte;
fopt=8 ('entab') counted it as two columns and stepped two bytes, so the
third byte was emitted on its own. Either way the file got invalid
UTF-8, on :r and :w alike. The same vcol += 2 also miscounted a three-
or four-byte character's columns for fopt=4 ('expandtab'), giving a tab
after a kanji fewer spaces than its stop.

Every step is utf_lenat() bytes and every column count utf_width() now;
'replace' still writes its two halfwidth spaces and removes the rest of
the character with a memmove.

:set fopt= / :set option= were also stored through *(long *) even though
p_opt and b_p_opt are int -- an eight-byte misaligned store over a
four-byte field, which UndefinedBehaviorSanitizer reports the first
time a test sets the option; it is *(int *) now.

+++ src/param.c
@@ -847,7 +847,7 @@ doset(char_u *arg)
 								|| (varp == (char_u *)&curbuf->b_p_opt))
 					{
 						char_u	*	p = arg + len + 1;
-						long		val = *(long *)(varp);
+						int			val = *(int *)(varp);
 
 						switch (*p) {
 						case '+':
@@ -860,7 +860,7 @@ doset(char_u *arg)
 							val = atol(p);
 							break;
 						}
-						*(long *)(varp) = val;
+						*(int *)(varp) = val;
 					}
 					else
 #endif
